requirements.txt
chromadb>=0.5.3
pandas>=2.2.2
openpyxl>=3.1.2
sentence-transformers>=3.0.1
pydantic<3
pyyaml>=6.0.1
